% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBtoLevel1.R
\name{allPerms}
\alias{allPerms}
\title{Generate permutations}
\usage{
allPerms(n)
}
\arguments{
\item{n}{size of permutations}
}
\value{
an n!xn matrix whose rows give permutations
}
\description{
Generate all permutations of 1 to n, as rows of a matrix
}
\examples{
allPerms(4)

}
