% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{ASTRALtreeLeopardusLescroart}
\alias{ASTRALtreeLeopardusLescroart}
\title{ASTRAL tree for Leopardus dataset}
\format{
an R data file
}
\usage{
data(ASTRALtreeLeopardusLescroart)
}
\description{
An .rda file containing the tree produced by ASTRAL from the "Leopardus" gene trees
of \insertCite{Lescroart2023;textual}{MSCquartets}. By the theory of \insertCite{ECTo2026}{MSCquartets}, 
this is a putative resolution of the tree of blobs for the unknown network, 
and can be used by \code{ECToBlob} to infer the network's tree of blobs.
}
\details{
If used for purposes other than learning software, please cite \insertRef{Lescroart2023}{MSCquartets}.
}
\references{
\insertRef{Lescroart2023}{MSCquartets}, \insertRef{ECTo2026}{MSCquartets}
}
\keyword{datasets}
