% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeOfBlobs.R
\name{collapseEdges}
\alias{collapseEdges}
\title{Collapse short tree edges}
\usage{
collapseEdges(tree, delta)
}
\arguments{
\item{tree}{a phylo object}

\item{delta}{minimum edge length to retain}
}
\value{
a phylo object
}
\description{
Set all edges of a tree with short lengths to be zero.
}
\examples{
tree=read.tree(text="((a:1,b:1):.5,(c:.5,d:2):1);")
newtree=collapseEdges(tree,delta=1)
write.tree(newtree)
}
