% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECToBlob.R
\name{combineP}
\alias{combineP}
\title{Combine p-values of tests}
\usage{
combineP(pvec, method = "Bon")
}
\arguments{
\item{pvec}{a vector of p-values to be combined}

\item{method}{the combination method to be used, one of \code{"Cauchy"}, \code{"Bon"} (default), \code{"BBC"}, \code{"CBC"}, or \code{"all"}}
}
\value{
If the argument \code{method} is \code{"Cauchy"}, \code{"Bon"}, \code{"BBC"}, or \code{"CBC"} a single p-value obtained from that method is returned.
If \code{method} is "All", a vector of all 4 are returned, with entries names "p_Cauchy", "p_Bon", "p_BBC", and "p_CBC"
}
\description{
Given a vector of p-values from not-necessarily-independent tests, compute a combined p-value
using the Cauchy combination method Liu and Xie (2020), the Bonferroni method (also called MinP, the minimum times the number of tests),
or the combined Cauchy/Bon variants BBC and CBC as proposed by Chen (2022).
}
\details{
All \code{NA}a in vector are ignored.
}
\examples{
combineP(c(.01,.04,.11))
combineP(c(.01,.04,.11),"all")

}
\references{
\insertRef{LiuXie2020}{MSCquartets}

\insertRef{Chen2022}{MSCquartets}
}
