% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeOfBlobs.R
\name{cutDensity}
\alias{cutDensity}
\title{Probability density function for Cut Model}
\usage{
cutDensity(lambda, mu0, alpha0, beta0)
}
\arguments{
\item{lambda}{statistic value (e.g., likelihood ratio statistic, or other power divergence statistic)}

\item{mu0}{parameter of density function}

\item{alpha0}{parameter of density function}

\item{beta0}{parameter of density function}
}
\value{
value of density function
}
\description{
Value of probability density function for Cut Model of \insertCite{ABMR24;textual}{MSCquartets}, Section 3.
}
\references{
\insertRef{ABMR24}{MSCquartets}
}
\seealso{
\code{\link{T1density},\link{T3density}}
}
