% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets.R
\docType{data}
\name{dataGeneTreeSample}
\alias{dataGeneTreeSample}
\title{Simulated gene tree dataset from species tree}
\format{
A text file with 1000 metric Newick gene trees on the taxa t1-t9
}
\description{
A text file dataset containing 1000 gene trees on 9 taxa simulated under the MSC on a species tree
}
\details{
This simulated dataset was produced by SimPhy \insertCite{simphy}{MSCquartets}, using the species tree

((((t5:5000,t6:5000):5000,t4:10000):2500,t7:12500):7500,((t8:3000,t9:3000):5000,\cr
((t1:4000,t2:4000):2500,t3:6500):1500):12000);

with a population size of 10,000 throughout the tree.

File is accessed as  \code{system.file("extdata","dataGeneTreeSample",package="MSCquartets")}, for example
via the ape command:

\code{gts=read.tree(file = system.file("extdata","dataGeneTreeSample",package="MSCquartets") )}
}
\references{
\insertRef{simphy}{MSCquartets}
}
