% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{dataHeliconiusMartin}
\alias{dataHeliconiusMartin}
\title{Heliconius gene tree dataset}
\format{
A text file with 2909 metric Newick gene trees each with 16 leaves labelled:\cr
chioneus.553, chioneus.560, chioneus.564, chioneus.565, \cr
ethilla.67, hecale.273, melpomeneFG.13435, melpomeneFG.9315, \cr
melpomeneFG.9316, melpomeneFG.9317, pardalinus.371, rosina.2071, \cr
rosina.531, rosina.533, rosina.546, sergestus.202
}
\source{
\doi{10.5061/dryad.dk712}
}
\description{
A text file dataset for Heliconius butterflies containing 2909 gene trees on 7 taxa, with 4 individuals sampled for each of 3 of the taxa, for a total of 16
leaves per gene tree. This is a subset of the data of \insertCite{Martin2013;textual}{MSCquartets}.
}
\details{
File is accessed as \code{system.file("extdata","dataHeliconiusMartin",package="MSCquartets")}, for
example via the \code{ape} command:

\code{gts = read.tree(file=system.file("extdata","dataHeliconiusMartin",package="MSCquartets"))}
}
\references{
\insertRef{Martin2013}{MSCquartets}
}
