% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{dataPapioniniVanderpool}
\alias{dataPapioniniVanderpool}
\title{Papionini gene tree dataset}
\format{
A text file with 1703 metric Newick gene trees each with 7 leaves labelled:\cr
Cercocebus_atys, Mandrillus_leucophaeus, Papio_anubis, Theropithecus_gelada,
Macaca_fascicularis, Macaca_mulatta, Macaca_nemestrina
}
\source{
\doi{10.5061/dryad.rfj6q577d}
}
\description{
A text file dataset for Papionini containing 1730 gene trees on 7 taxa. This is a subset of the data of \insertCite{Vanderpool2020;textual}{MSCquartets}.
}
\details{
File is accessed as \code{system.file("extdata","dataPapioniniVanderpool",package="MSCquartets")}, for
example via the \code{ape} command:

\code{gts = read.tree(file=system.file("extdata","dataPapioniniVanderpool",package="MSCquartets"))}
}
\references{
\insertRef{Vanderpool2020}{MSCquartets}
}
