% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{dataYeastRokas}
\alias{dataYeastRokas}
\title{Yeast gene tree dataset}
\format{
A text file with 106 topological Newick gene trees on the taxa:
 Sbay, Scas, Scer, Sklu, Skud, Smik, Spar, and Calb (outgroup).
}
\source{
from L. Nakhleh's website at Rice U., link no longer available as of 2026
}
\description{
A text file dataset for Yeast containing 106 gene trees on 8 taxa  (7 Saccharomyces and 1 Candida outgroup). This is a subset of the data
of \insertCite{Rokas03;textual}{MSCquartets}.
}
\details{
File is accessed as \code{system.file("extdata","dataYeastRokas",package="MSCquartets")}, for example
via the \code{ape} command:

\code{ gts=read.tree(file = system.file("extdata","dataYeastRokas",package="MSCquartets"))}
}
\references{
\insertRef{Rokas03}{MSCquartets}
}
