% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBtoLevel1.R
\name{expmodNANUQCycleDist}
\alias{expmodNANUQCycleDist}
\title{Expected modNANUQ cycle distance}
\usage{
expmodNANUQCycleDist(n)
}
\arguments{
\item{n}{number of edges in cycle (at least 4)}
}
\value{
an nxn distance matrix with rows/columns ordered from hybrid following circular order
}
\description{
Compute expected modNANUQ distance for a sunlet network with 4 or more taxa.
This is used in a hueristic method for resolving multifurcation in a tree
of blobs to a cycle by NANUQ+ commands.
}
\examples{
expmodNANUQCycleDist(5)

#'@seealso \code{\link{resolveCycle}} \code{\link{ordersHeuristicmodNANUQ}}

}
