% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBtoLevel1.R
\name{expNANUQCycleDist}
\alias{expNANUQCycleDist}
\title{Expected NANUQ cycle distance}
\usage{
expNANUQCycleDist(n)
}
\arguments{
\item{n}{number of edges in cycle}
}
\value{
an nxn distance matrix with rows/columns ordered from hybrid following circular order
}
\description{
Compute expected NANUQ distance for a sunlet network with 4 or more taxa. This is used
to resolve multifurcations in a tree of blobs by NANUQ+ functions
}
\examples{
expNANUQCycleDist(5)

}
\seealso{
\code{\link{resolveCycle}}
}
