% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeQuartetHypTest.R
\name{M0}
\alias{M0}
\title{Modified Struve function}
\usage{
M0(x)
}
\arguments{
\item{x}{function argument}
}
\value{
value of negative modified Struve function function
}
\description{
This function is used in computing the probability density for
Model T1. The code is closely based on the \code{I0L0} function implemented in Python for the package 
RandomFieldUtils, which was previously on CRAN up to 12/2022).
}
