% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBtoLevel1.R
\name{ordersHeuristicmodNANUQ}
\alias{ordersHeuristicmodNANUQ}
\title{Choose cycle orders heuristically from  empirical modNANUQ distance}
\usage{
ordersHeuristicmodNANUQ(M, delta = 10^-6)
}
\arguments{
\item{M}{an empirical modNANUQ distance table}

\item{delta}{cutoff for relative difference in distances for determining near ties for "best" orders}
}
\value{
a list of circular orders
}
\description{
Find candidates for best hybrid node and circular order fitting the modNANUQ
distance.
}
\details{
Candidadte orders are obtained  by first picking the hybrid node (from the
minimum column sum of the distance matrix), then ordering nodes by distance
from the hybrid, and for each consecutive pair picking  nodes in the cycle
closest to the previous node. This constructs one or more orders since ties
may occur. For more details, see \insertCite{ABRW24;textual}{MSCquartets}.

This function is used by NANUQ+ commands
to resolve multifurcations in a tree of blobs of high degree.
}
\references{
\insertRef{ABMR24}{MSCquartets}
}
\seealso{
\code{\link{expmodNANUQCycleDist}} \code{\link{resolveCycle}} \code{\link{resolveLevel1}}
}
