% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeQuartetHypTest.R
\name{powerDivStat}
\alias{powerDivStat}
\title{Power divergence statistic of Cressie & Read}
\usage{
powerDivStat(obs, expd, lambda)
}
\arguments{
\item{obs}{observation vector}

\item{expd}{expected vector}

\item{lambda}{statistic parameter (e.g., 0=Likelihood Ratio, 1=Chi-squared)}
}
\value{
value of statistic
}
\description{
Computes any of the family of power-divergence statistics for multinomial data of \insertCite{CR84;textual}{MSCquartets},
to compare observed and expected counts. Includes Likelihood Ratio and Chi-squared statistics as special cases.
}
\examples{
obs=c(10,20,30)
expd=c(20,20,20)
powerDivStat(obs,expd,0)

}
\references{
\insertRef{CR84}{MSCquartets}
}
