% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECToBlob.R
\name{quartetECtestInd}
\alias{quartetECtestInd}
\title{Hypothesis testing of quartet counts for ECToBlob}
\usage{
quartetECtestInd(
  rqt,
  tree,
  lambda = 0,
  method = "MLest",
  smallsample = "precomputed",
  smallcounts = "precomputed",
  bootstraps = 10^4
)
}
\arguments{
\item{rqt}{either a table of resolved quartet counts from a collection of gene trees,
or that table with additional columns of quartet p-values.}

\item{tree}{an unrooted tree determining the topologies of the displayed quartet trees to test}

\item{lambda}{parameter for power-divergence statistic (e.g., 0 for likelihood ratio statistic, 1 for Chi-squared statistic)}

\item{method}{"MLest", "conservative", or "bootstrap"}

\item{smallsample}{"precomputed" or "bootstrap", method of obtaining p-value when sample is small (<30)}

\item{smallcounts}{"precomputed" or "bootstrap", method of obtaining p-value when some (but not all) counts are small}

\item{bootstraps}{number of samples for bootstrapping}
}
\value{
the same table as \code{rqt} with three columns appended:
\code{qindex-D} containing the indicies (1,2,or 3)
of the quartet topology displayed on \code{tree}, \code{p_T1-D}
containing the p-value for the T1 test of the displayed quartet topologies.
\code{p_star} containing p-values for the star tree test.
}
\description{
Append p-values for the star and T1 tests of displayed quartets on a given
tree to resolved quartet count table. This is used in \code{ECToBlob}
to obtain an estimated Tree of Blobs for a data set from a supplied tree that
is assumed to be a resolution of the Tree of Blobs. Three columns are appended
to the quartet count table: "p_T1-D" holds the T1 test p-values, "qindexD" holds
the index (1,2,or 3) of the quartet topology displayed on the tree, and "p_star"
holds the star tree test p-value.
}
\details{
If on the tree a quartet is unresolved, p_T1 and qindex column entry will be NA

\code{tree} must be supplied as phylo object, but may have multifurcations.

If either p_star or p_T1-D and qindex-D values are already in table, they are
not recomputed.
}
\examples{
data(ASTRALtreeLeopardusLescroart)
Atree=read.tree(text=ASTRALtreeLeopardusLescroart)
data(tableLeopardusLescroart)
pTable=quartetECtestInd(tableLeopardusLescroart,Atree)

}
