% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarQuartetHypTest.R
\name{quartetStarTest}
\alias{quartetStarTest}
\title{Hypothesis test for quartet counts fitting a star tree under the MSC}
\usage{
quartetStarTest(obs)
}
\arguments{
\item{obs}{vector of 3 counts of resolved quartet frequencies}
}
\value{
p-value
}
\description{
Perform hypothesis test for star tree for a vector of  quartet counts to fit expected frequencies of (1/3,1/3,1/3).
The test performed is a standard Chi-square with 2 degrees of freedom.
}
\examples{
obs=c(16,72,12)
quartetStarTest(obs)

}
