% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexPlot.R
\name{simplexLabels}
\alias{simplexLabels}
\title{Label vertices of 2-d probability simplex}
\usage{
simplexLabels(top = "", left = "", right = "")
}
\arguments{
\item{top}{label for top}

\item{left}{label for left bottom}

\item{right}{label for right bottom}
}
\value{
No return value, called for side effects
}
\description{
Add labels to vertices of the probability simplex.
}
\examples{
   simplexPrepare("T3","Example Plot")
   simplexLabels("ab|cd","ac|bd","ad|bc")

}
\seealso{
\code{\link{simplexPoint}},
         \code{\link{simplexPrepare}},
         \code{\link{simplexSegment}},
         \code{\link{simplexText}},
         \code{\link{simplexCoords}}
}
