% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexPlot.R
\name{simplexPoint}
\alias{simplexPoint}
\title{Plot point in 2-d probability simplex}
\usage{
simplexPoint(v, ...)
}
\arguments{
\item{v}{a 3-d point in non-negative orthant, coordinates not summing to 0}

\item{...}{other options to pass to graphics::points function}
}
\value{
No return value, called for side effects
}
\description{
Normalizes a point given in 3-d non-normalized coordinates, then plots it in 
the 2-d probability simplex.
}
\examples{
   simplexPrepare("T3","Example Plot")
   simplexPoint(c(15,65,20),pch=3,col="blue")

}
\seealso{
\code{\link{simplexLabels}},
         \code{\link{simplexPrepare}},
         \code{\link{simplexSegment}},
         \code{\link{simplexText}},
         \code{\link{simplexCoords}}
}
