% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexPlot.R
\name{simplexPrepare}
\alias{simplexPrepare}
\title{Draw 2-d probability simplex, with model lines for T3 or T1 model}
\usage{
simplexPrepare(model = "T3", maintitle = NULL, titletext = NULL)
}
\arguments{
\item{model}{\code{"T1"} or \code{"T3"}, for 1-tree or 3-tree model}

\item{maintitle}{main title for plot}

\item{titletext}{additional text for title}
}
\description{
Outline the 2-d probability simplex, and draw the T1 or T3 model points for quartet frequencies.
The models "T1" and "T3" are described more fully by \insertCite{MAR19;textual}{MSCquartets}.
}
\examples{
   simplexPrepare("T3",maintitle="Main title",titletext="further text")

}
\references{
\insertRef{MAR19}{MSCquartets}
}
\seealso{
\code{\link{simplexLabels}},
         \code{\link{simplexPoint}},
         \code{\link{simplexSegment}},
         \code{\link{simplexText}},
         \code{\link{simplexCoords}}
}
