% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexPlot.R
\name{simplexSegment}
\alias{simplexSegment}
\title{Plot line segment in 2-d probability simplex}
\usage{
simplexSegment(v, w, ...)
}
\arguments{
\item{v, w}{3-d endpoints of line segment in non-negative orthant, coords not summing to 0}

\item{...}{other options to pass to graphics::segments function}
}
\value{
No return value, called for side effects
}
\description{
Normalizes two points in 3-d, and draws line segment between them in 2-d probability simplex.
}
\examples{
   simplexPrepare("T3","Example Plot")
   simplexSegment(c(15,65,20),c(15,70, 15),col="green")

}
\seealso{
\code{\link{simplexLabels}},
         \code{\link{simplexPoint}},
         \code{\link{simplexPrepare}},
         \code{\link{simplexText}},
         \code{\link{simplexCoords}}
}
