% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplexPlot.R
\name{simplexText}
\alias{simplexText}
\title{Add text at a point in 2-d probability simplex}
\usage{
simplexText(v, label = "", ...)
}
\arguments{
\item{v}{a 3-d point in non-negative orthant, coordinates not summing to 0}

\item{label}{text to add to plot}

\item{...}{other options to pass to graphics::text function}
}
\value{
No return value, called for side effects
}
\description{
Add text to a 2-d probability simplex plot, at specified location.
}
\examples{
   simplexPrepare("T3","Example Plot")
   simplexText(c(15,65,20),"tree ac|bd")

}
\seealso{
\code{\link{simplexLabels}},
         \code{\link{simplexPoint}},
         \code{\link{simplexPrepare}},
         \code{\link{simplexSegment}},
         \code{\link{simplexCoords}}
}
