% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetTally.R
\name{sortQuartetTableRows}
\alias{sortQuartetTableRows}
\title{Sort quartet table rows}
\usage{
sortQuartetTableRows(qT)
}
\arguments{
\item{qT}{a quartet Table to be sorted}
}
\value{
sorted table
}
\description{
Sort the rows of a quartet table so they are in MSCquartet's standard lex order.
This is the order produced by the \code{quartetTable} function. The only exceptions
to this order produced in the package are when \code{quartetTable} is called with the
\code{random} argument non-zero, or when the \code{HolmBonferoni} function is called.
However, for tables made outside this package, it can be useful.
}
\seealso{
\code{\link{quartetTable}}, \code{\link{HolmBonferroni}}
}
