% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evonetMisc.R
\name{suppressNetBinaryNodes}
\alias{suppressNetBinaryNodes}
\title{Suppress Network Binary Nodes}
\usage{
suppressNetBinaryNodes(net, addlength = TRUE)
}
\arguments{
\item{net}{a phylogenetic network, of class "evonet"}

\item{addlength}{if TRUE (default), add lengths of two incident edges for new edge.}
}
\description{
Suppress any non-root binary nodes in a phylogenetic network.
}
\details{
This function is similar to ape's \code{collapse.singles} which only works
on phylo objects that are trees.
}
