% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeQuartetHypTest.R
\name{T1density}
\alias{T1density}
\title{Probability density function for Model T1 of \insertCite{MAR19;textual}{MSCquartets}, Proposition 5.2}
\usage{
T1density(x, mu0)
}
\arguments{
\item{x}{statistic value (e.g., likelihood ratio statistic, or other power divergence statistic);
may be a vector}

\item{mu0}{parameter of density function}
}
\value{
value of density function
}
\description{
Value of probability density function for Model T1 of \insertCite{MAR19;textual}{MSCquartets}, Proposition 5.2.
}
\references{
\insertRef{MAR19}{MSCquartets}
}
\seealso{
\code{\link{T3density}}
}
