% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeQuartetHypTest.R
\name{T3density}
\alias{T3density}
\title{Probability density function for Model T3}
\usage{
T3density(x, mu0, alpha0, beta0)
}
\arguments{
\item{x}{statistic value (e.g., likelihood ratio statistic, or other power divergence statistic)}

\item{mu0}{parameter of density function}

\item{alpha0}{parameter of density function}

\item{beta0}{parameter of density function}
}
\value{
value of density function
}
\description{
Value of probability density function for Model T3 of \insertCite{MAR19;textual}{MSCquartets}, Proposition 4.2.
}
\references{
\insertRef{MAR19}{MSCquartets}
}
\seealso{
\code{\link{T1density}}
}
