% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{tableHeliconiusMartin}
\alias{tableHeliconiusMartin}
\alias{tableHeleconiusMartin}
\title{Quartet table for Heliconius gene tree dataset}
\format{
an R data file
}
\source{
\doi{10.5061/dryad.dk712}
}
\usage{
data(tableHeliconiusMartin)
}
\description{
An .rda file containing a quartet table summarizing the "Heleconius" gene trees
of \insertCite{Martin2013;textual}{MSCquartets}. This table contains quartet counts
from 2909 gene trees on 7 taxa, with 4 individuals sampled for each of 3 of the taxa,
for a total of 16 leaves per gene tree.
}
\details{
This table is provided rather than the original gene trees to save storage space.
If used, please cite \insertCite{Martin2013;textual}{MSCquartets}.
}
\references{
\insertRef{Martin2013}{MSCquartets}
}
\keyword{datasets}
