% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets-package.R
\docType{data}
\name{tableLeopardusLescroart}
\alias{tableLeopardusLescroart}
\title{Quartet table for Leopardus dataset}
\format{
an R data file
}
\usage{
data(tableLeopardusLescroart)
}
\description{
An .rda file containing a quartet table summarizing the "Leopardus" gene trees
of \insertCite{Lescroart2023;textual}{MSCquartets}. This table contains quartet counts
for 16,338 "gene" trees for 16 taxa, inferred from sliding widows across
genomic sequences.
}
\details{
This table is provided rather than the original gene trees to save storage space.
If used, please cite \insertRef{Lescroart2023}{MSCquartets}.
}
\references{
\insertRef{Lescroart2023}{MSCquartets}
}
\keyword{datasets}
