% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetTally.R
\name{taxonNames}
\alias{taxonNames}
\title{Get all taxon names from a collection of trees}
\usage{
taxonNames(trees)
}
\arguments{
\item{trees}{a multiPhylo object containing a collection of trees}
}
\value{
a vector of unique names of taxa appearing on the trees
}
\description{
Create a vector of all taxon names appearing on a collection of trees, with no repeats.
}
\examples{
gtrees=read.tree(file=system.file("extdata","dataGeneTreeSample",package="MSCquartets"))
tnames=taxonNames(gtrees)

}
