% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCquartets.R
\docType{data}
\name{dataYeastRokas}
\alias{dataYeastRokas}
\title{Yeast gene tree dataset}
\format{
A text file with 106 topological Newick gene trees on the taxa: 
 Sbay, Scas, Scer, Sklu, Skud, Smik, Spar, and Calb (outgroup).
}
\source{
\url{https://wiki.rice.edu/confluence/download/attachments/8898533/yeast.trees?version=1&modificationDate=1360603275797&api=v2}
}
\description{
A text file dataset for Yeast containing 106 gene trees on 8 taxa  (7 Saccharomyces and 1 Candida outgroup). This is a subset of the data
of \insertCite{Rokas03;textual}{MSCquartets}.
}
\details{
File is accessed as \code{system.file("extdata","dataYeastRokas",package="MSCquartets")}, for example
via the \code{ape} command:

\code{ gts=read.tree(file = system.file("extdata","dataYeastRokas",package="MSCquartets"))}
}
\references{
\insertRef{Rokas03}{MSCquartets}
}
