% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initBquartets}
\alias{initBquartets}
\title{Initialize vector of B quartets}
\usage{
initBquartets(pTable, m, alpha, beta, colptest, colpstar, Bquartets)
}
\arguments{
\item{pTable}{a quartet table with p-values}

\item{m}{number of rows in pTable}

\item{alpha}{critical value for tree test}

\item{beta}{critical value for star tree test}

\item{colptest}{column with p value for tree test}

\item{colpstar}{column with p value for star tree test}

\item{Bquartets}{0/1 vector encoding initial B quartets}
}
\value{
two lists, "Lout" with indices of initial B-quartets, "Bout" with TRUE/FALSE for all sets
of 4 taxa indicating which are B-quartets
}
\description{
This is a C++ function, used in \code{TINNIKdist}, to initialize for
inference of B and T quartets. Doucumentation if for internal use only.
}
\seealso{
\code{\link{quartetTable}}, \code{\link{quartetTableParallel}}
}
