% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quartetTallyCpp}
\alias{quartetTallyCpp}
\title{Build quartet table from distances}
\usage{
quartetTallyCpp(dList, M, nt, Q, random)
}
\arguments{
\item{dList}{a list of distance matrices}

\item{M}{number of sets of 4 taxa}

\item{nt}{number of gene trees/distance matrices}

\item{Q}{matrix to fill out as table of quartet counts}

\item{random}{if 0 compute for all sets of 4 taxa, otherwise for M random ones}
}
\description{
This is a C++ function, called from quartetTable, to fill in the quartet counts.
From a list of topological distance matrices (1 for each gene tree) it determines all
gene quartets. It is not intended to be used as a stand-alone function, and hence not fully
documented. The faster looping in C++ over R gives substantial time improvements
}
\seealso{
\code{\link{quartetTable}}, \code{\link{quartetTableParallel}}
}
